# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mobin AI API",
    "description": "Scoped 100-question GraphRAG API. Register, exchange the one-time grant, then authorize with the returned bearer token.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/clients": {
      "post": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Register a visitor",
        "description": "Verifies Turnstile before storing contact information. Safe to retry with a fresh Turnstile response; a grant can be exchanged once.",
        "operationId": "register_api_v1_clients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOut"
                }
              }
            }
          },
          "400": {
            "description": "Turnstile verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "IP registration rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations": {
      "post": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Exchange a one-time registration grant",
        "description": "The client_id alone is not a credential. The grant expires after five minutes and cannot issue another token. Each token has 100 chargeable answers and a 30-day TTL.",
        "operationId": "create_conversation_api_v1_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationOut"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or used grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conversation already issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Conversation Info",
        "operationId": "conversation_info_api_v1_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationInfo"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Token is scoped to another conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{conversation_id}/messages": {
      "get": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Messages",
        "operationId": "messages_api_v1_conversations__conversation_id__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Token is scoped to another conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Post Message",
        "description": "Idempotency-Key is required. A completed replay returns the stored outcome without another quota charge. Completed and insufficient-evidence answers consume one of 100; failures do not.",
        "operationId": "post_message_api_v1_conversations__conversation_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 160,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Token is scoped to another conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "In-flight generation or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Quota, rate, or model budget exceeded; inspect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Knowledge or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{conversation_id}/messages:stream": {
      "post": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Stream Message",
        "operationId": "stream_message_api_v1_conversations__conversation_id__messages_stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 160,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "POST SSE: retrieving, citation, delta, complete, error. The complete event follows durable persistence.",
            "content": {
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Token is scoped to another conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "In-flight generation or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Quota, rate, or model budget exceeded; inspect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Knowledge or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/feedback": {
      "post": {
        "tags": [
          "Mobin AI"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_v1_messages__message_id__feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Token is scoped to another conversation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Answer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationIn": {
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "registration_grant": {
            "type": "string",
            "minLength": 20,
            "title": "Registration Grant"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "registration_grant"
        ],
        "title": "ConversationIn",
        "examples": [
          {
            "client_id": "01900000-0000-7000-8000-000000000000",
            "registration_grant": "one-time-grant-from-registration"
          }
        ]
      },
      "ConversationInfo": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "quota": {
            "$ref": "#/components/schemas/Quota"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "expires_at",
          "quota"
        ],
        "title": "ConversationInfo"
      },
      "ConversationOut": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "quota": {
            "$ref": "#/components/schemas/Quota"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "access_token",
          "expires_at",
          "quota"
        ],
        "title": "ConversationOut",
        "examples": [
          {
            "access_token": "opaque-random-token",
            "conversation_id": "01900000-0000-7000-8000-000000000001",
            "expires_at": "2026-10-25T12:00:00Z",
            "quota": {
              "limit": 100,
              "remaining": 100,
              "used": 0
            }
          }
        ]
      },
      "ErrorOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorOut"
      },
      "FeedbackIn": {
        "properties": {
          "helpful": {
            "type": "boolean",
            "title": "Helpful"
          }
        },
        "type": "object",
        "required": [
          "helpful"
        ],
        "title": "FeedbackIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MessagePage"
      },
      "QuestionIn": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QuestionIn",
        "examples": [
          {
            "question": "How should a high-TPS Kafka to ClickHouse pipeline be designed?"
          }
        ]
      },
      "Quota": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "used",
          "remaining"
        ],
        "title": "Quota"
      },
      "RegisterIn": {
        "properties": {
          "given_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Given Name"
          },
          "family_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Family Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "privacy_accepted": {
            "type": "boolean",
            "title": "Privacy Accepted"
          },
          "privacy_policy_version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Privacy Policy Version"
          },
          "marketing_consent": {
            "type": "boolean",
            "title": "Marketing Consent",
            "default": false
          },
          "turnstile_token": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Turnstile Token"
          }
        },
        "type": "object",
        "required": [
          "given_name",
          "family_name",
          "privacy_accepted",
          "privacy_policy_version",
          "turnstile_token"
        ],
        "title": "RegisterIn",
        "examples": [
          {
            "email": "ada@example.com",
            "family_name": "Lovelace",
            "given_name": "Ada",
            "marketing_consent": false,
            "privacy_accepted": true,
            "privacy_policy_version": "2026-09-25",
            "turnstile_token": "turnstile-response"
          }
        ]
      },
      "RegisterOut": {
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "registration_grant": {
            "type": "string",
            "title": "Registration Grant"
          },
          "grant_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Grant Expires At"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "registration_grant",
          "grant_expires_at"
        ],
        "title": "RegisterOut",
        "examples": [
          {
            "client_id": "01900000-0000-7000-8000-000000000000",
            "grant_expires_at": "2026-09-25T12:05:00Z",
            "registration_grant": "short-lived-one-time-grant"
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}